Move `UploadResultStatus` and `UploadResult` to upload/uploader.ts
This removes the need of an import from the uploader interface
file and breaks a deep circular dependency on Glean.ts in the
process.
