+++ glean/src/platform/webext/uploader.ts
@@ -2,8 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import Uploader, { DEFAULT_UPLOAD_TIMEOUT_MS } from "../../core/upload/uploader.js";
-import { UploadResult, UploadResultStatus } from "../../core/upload/index.js";
+import Uploader, { DEFAULT_UPLOAD_TIMEOUT_MS, UploadResult, UploadResultStatus } from "../../core/upload/uploader.js";
 
 class BrowserUploader implements Uploader {
   async post(url: string, body: string, headers: Record<string, string> = {}): Promise<UploadResult> {
