+++ glean/tests/platform/uploader.spec.ts
@@ -6,9 +6,9 @@ import "jsdom-global/register";
 import assert from "assert";
 import sinon from "sinon";
 
-import { UploadResultStatus } from "../../src/core/upload";
 import BrowserUploader from "../../src/platform/webext/uploader";
 import { JSONObject } from "../../src/core/utils";
+import { UploadResultStatus } from "../../src/core/upload/uploader";
 
 const sandbox = sinon.createSandbox();
